# Performance: Phase 1 startup optimizations - 76% improvement with dynamic imports